fix: enable X11 and Wayland features for eframe on Linux
The winit compilation errors were caused by eframe having
default-features = false, which disabled winit's platform support.

On Linux, winit requires either X11 or Wayland features to be
explicitly enabled when default features are disabled.

Added 'x11' and 'wayland' to eframe's features list to fix
compilation on Linux when using --all-features (which is used
in CI tests and clippy checks).

Error: 'The platform you're compiling for is not supported by winit'